+++ Assets/Scripts/StateMachine/Prepare_ApplicationState.cs
@@ -57,7 +57,7 @@ public class Prepare_ApplicationState : ApplicationState
         //Debug.Log("state" + m_type + " ending! ");
     }
 
-    public override int GetMainThemeSchedule()
+    public override float GetMainThemeSchedule()
     {
         //return 86;
         return 279;
